Successfully added in the guard function to the bot.
Guard is a concept based on middleware as well the word originated from Nest.JS. The point of the guard is to make sure that the user has been given the permission to use certain functions of the bot. Example, the leveling system has a blacklist system in place. If the user is blacklisted, then they can't get exp, so a guard was made to check the blacklist to see if the user is in that list or not.